requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
requests==2.31.0
service_streamer==0.1.2
fastapi_utils==0.2.1
pillow==10.2.0
python-multipart==0.0.9
python-dotenv==1.0.1
websocket-client==1.7.0
opencv-python==4.9.0.80
